+++ src/core/shared/interfaces/TextLabel.ts
@@ -3,4 +3,5 @@ import { Vector3 } from './Vector';
 export interface TextLabel {
     pos: Vector3;
     data: string;
+    maxDistance?: number;
 }